backend/pyproject.toml
[project]
name = "filesh-backend"
version = "0.1.0"
description = "Backend service for the file sharing application"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.15.2",
  "fastapi>=0.115.12",
  "pydantic-settings>=2.8.1",
  "python-json-logger>=2.0.7",
  "sqlalchemy>=2.0.40",
  "psycopg[binary]>=3.2.6",
  "uvicorn[standard]>=0.34.1",
  "pyjwt>=2.12.1",
  "email-validator>=2.3.0",
  "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pre-commit>=4.5.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.1.1",
  "ruff>=0.11.5",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]
omit = ["app/**/__init__.py"]

[tool.alembic]
script_location = "alembic"
